pyproject.toml
[project]
name = "opeelcky"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huey>=2.5.3",
    "ipykernel>=6.29.5",
    "ipython>=9.1.0",
    "jsonpath-ng>=1.7.0",
    "pandas>=2.2.3",
    "pip>=25.0.1",
    "psycopg2-binary>=2.9.10",
    "pudb>=2024.1.3",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "redis>=5.2.1",
    "requests>=2.32.3",
]
